Refactor Runners, introduce `Task` class (#4206)
* Introduce task class + small refactoring
* Merge in main and make release_datasets private
* Fix session tests
* Make Task runnable and call inside runners
* Move helper methods from parallel runner to task
* Mark run_node as deprecated
* Refactor helper methods to go inside Task, making hook_manager an optional argument, and adding parallel as boolean flag

---------
