+++ kedro/runner/__init__.py
@@ -5,12 +5,14 @@ to execute ``Pipeline`` instances.
 from .parallel_runner import ParallelRunner
 from .runner import AbstractRunner, run_node
 from .sequential_runner import SequentialRunner
+from .task import Task
 from .thread_runner import ThreadRunner
 
 __all__ = [
     "AbstractRunner",
     "ParallelRunner",
     "SequentialRunner",
+    "Task",
     "ThreadRunner",
     "run_node",
 ]
