+++ kedro/runner/sequential_runner.py
@@ -9,7 +9,8 @@ from collections import Counter
 from itertools import chain
 from typing import TYPE_CHECKING, Any
 
-from kedro.runner.runner import AbstractRunner, run_node
+from kedro.runner.runner import AbstractRunner
+from kedro.runner.task import Task
 
 if TYPE_CHECKING:
     from pluggy import PluginManager
@@ -75,21 +76,20 @@ class SequentialRunner(AbstractRunner):
 
         for exec_index, node in enumerate(nodes):
             try:
-                run_node(node, catalog, hook_manager, self._is_async, session_id)
+                Task(
+                    node=node,
+                    catalog=catalog,
+                    hook_manager=hook_manager,
+                    is_async=self._is_async,
+                    session_id=session_id,
+                ).execute()
                 done_nodes.add(node)
             except Exception:
                 self._suggest_resume_scenario(pipeline, done_nodes, catalog)
                 raise
 
-            # decrement load counts and release any datasets we've finished with
-            for dataset in node.inputs:
-                load_counts[dataset] -= 1
-                if load_counts[dataset] < 1 and dataset not in pipeline.inputs():
-                    catalog.release(dataset)
-            for dataset in node.outputs:
-                if load_counts[dataset] < 1 and dataset not in pipeline.outputs():
-                    catalog.release(dataset)
+            self._release_datasets(node, catalog, load_counts, pipeline)
 
             self._logger.info(
-                "Completed %d out of %d tasks", exec_index + 1, len(nodes)
+                "Completed %d out of %d tasks", len(done_nodes), len(nodes)
             )
