+++ pyproject.toml
@@ -172,7 +172,7 @@ layers = [
     "config"
 ]
 ignore_imports = [
-    "kedro.runner.parallel_runner -> kedro.framework.project",
+    "kedro.runner.task -> kedro.framework.project",
     "kedro.framework.hooks.specs -> kedro.framework.context"
 ]
