+++ tests/framework/session/test_session_extension_hooks.py
@@ -22,7 +22,7 @@ from kedro.io import DataCatalog, MemoryDataset
 from kedro.pipeline import node, pipeline
 from kedro.pipeline.node import Node
 from kedro.runner import ParallelRunner
-from kedro.runner.runner import _run_node_async
+from kedro.runner.task import Task
 from tests.framework.session.conftest import (
     _assert_hook_call_record_has_expected_parameters,
     _assert_pipeline_equal,
@@ -580,11 +580,13 @@ class TestAsyncNodeDatasetHooks:
         mock_session.load_context()
 
         # run the node asynchronously with an instance of `LogCatalog`
-        _run_node_async(
+        task = Task(
             node=sample_node,
             catalog=memory_catalog,
             hook_manager=mock_session._hook_manager,
+            is_async=True,
         )
+        task.execute()
 
         hooks_log_messages = [r.message for r in logs_listener.logs]
 
@@ -604,11 +606,13 @@ class TestAsyncNodeDatasetHooks:
             hook_manager.hook, "after_dataset_saved"
         )
 
-        _run_node_async(
+        task = Task(
             node=sample_node_multiple_outputs,
             catalog=memory_catalog,
             hook_manager=hook_manager,
+            is_async=True,
         )
+        task.execute()
 
         after_dataset_saved_mock.assert_has_calls(
             [
